23/09

+++ Embedded/Robot_Raimbaud_Clement.X/PWM.h
@@ -5,7 +5,7 @@
 #define MOTEUR_GAUCHE 1
 
 void InitPWM(void);
-void PWMSetSpeedConsigne(int, float);
+void PWMSetSpeedConsigne(float, int);
 void PWMUpdateSpeed(void);
 
 #endif /* PWM_H */
